# openapi.yaml
openapi: 3.0.3
info:
  title: Raptor API
  description: |
    Self-hosted webhook/email/DNS capture & inspection. This document is the
    spec-first source of truth for the versioned management API under `/api/v1`.
    The SPA is a pure client of it.

    **Authentication.** When `--require-auth` is enabled, send `Api-Key: <uuid>`.
    With no key configured the API is open (first-run bootstrap mode).
  version: "1.0"
servers:
  - url: /api/v1
paths:
...
  /tokens/{tokenID}/replay:
    parameters:
      - $ref: "#/components/parameters/TokenID"
    post:
      summary: Replay a subset of captured requests to a target URL
      operationId: replayRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_url]
              properties:
                target_url: { type: string }
                q: { type: string, description: "search DSL to select the subset" }
                date_from: { type: string }
                date_to: { type: string }
                limit: { type: integer }
      responses:
        "200":
          description: Replay counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  replayed: { type: integer }
                  failed: { type: integer }
...
  /tokens/{tokenID}/requests:
    parameters:
      - $ref: "#/components/parameters/TokenID"
    get:
      summary: List captured requests (paged, newest first)
      operationId: listRequests
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 50, maximum: 100 } }
        - name: q
          in: query
          description: >
            Search DSL. Free text matches content; field filters include
            `content:`, `method:`, `type:`, `ip:`, `host:`, `url:`, `query:`,
            `headers.<key>:`, `_exists_:<field>`, and
            `created_at:[* TO now-14d]`. Terms are ANDed; quote values with spaces.
          schema: { type: string }
        - { name: date_from, in: query, description: "ISO date/datetime lower bound", schema: { type: string } }
        - { name: date_to, in: query, description: "ISO date/datetime upper bound", schema: { type: string } }
      responses:
        "200":
          description: Page of requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Request" }
                  total: { type: integer }
                  page: { type: integer }
                  per_page: { type: integer }
    delete:
      summary: Delete requests for the token (all, or a `q`/`date_from`/`date_to` subset)
      operationId: deleteAllRequests
      parameters:
        - { name: q, in: query, description: "search DSL; deletes only matching requests", schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string } }
        - { name: date_to, in: query, schema: { type: string } }
      responses:
        "200":
          description: Number deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
...
  /tokens/{tokenID}/requests/latest:
    parameters:
      - $ref: "#/components/parameters/TokenID"
    get:
      summary: Get the most recent request
      operationId: latestRequest
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "404": { description: No requests yet }
  /tokens/{tokenID}/requests/{requestID}:
    parameters:
      - $ref: "#/components/parameters/TokenID"
      - $ref: "#/components/parameters/RequestID"
    get:
      summary: Get a single request
      operationId: getRequest
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "404": { description: Not found }
    delete:
      summary: Delete a single request
      operationId: deleteRequest
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
  /tokens/{tokenID}/requests/{requestID}/raw:
    parameters:
      - $ref: "#/components/parameters/TokenID"
      - $ref: "#/components/parameters/RequestID"
    get:
      summary: Get the raw HTTP representation of a request
      operationId: rawRequest
      responses:
        "200":
          description: Raw request
          content:
            text/plain:
              schema: { type: string }
  /tokens/{tokenID}/requests/{requestID}/files/{fileID}:
    parameters:
      - $ref: "#/components/parameters/TokenID"
      - $ref: "#/components/parameters/RequestID"
      - { name: fileID, in: path, required: true, schema: { type: string } }
    get:
      summary: Download a request's attached file
      operationId: downloadFile
      responses:
        "200":
          description: File blob
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404": { description: Not found }
components:
  parameters:
    TokenID:
      name: tokenID
      in: path
      required: true
      description: Token UUID or alias
      schema: { type: string }
    RequestID:
      name: requestID
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    Request:
      type: object
      properties:
        uuid: { type: string }
        token_id: { type: string }
        type: { type: string, enum: [web, email, dns] }
        method: { type: string }
        ip: { type: string }
        hostname: { type: string }
        user_agent: { type: string }
        content: { type: string }
        query: { type: object, additionalProperties: { type: array, items: { type: string } } }
        headers: { type: object, additionalProperties: { type: array, items: { type: string } } }
        url: { type: string }
        size: { type: integer }
        sorting: { type: integer, format: int64 }
        files:
          type: array
          items: { $ref: "#/components/schemas/File" }
        created_at: { type: string, format: date-time }
        # Email-specific fields (type == email)
        sender: { type: string }
        message_id: { type: string }
        destinations: { type: string }
        subject: { type: string }
        text_content: { type: string }
        checks:
          type: object
          description: "DKIM/SPF/DMARC results, e.g. {\"dkim\":\"pass\",\"spf\":\"fail\"}"
          additionalProperties: { type: string }
    File:
      type: object
      properties:
        id: { type: string }
        request_id: { type: string }
        filename: { type: string }
        content_type: { type: string }
        size: { type: integer }
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Api-Key